+++ src/PostgREST/Auth.hs
@@ -44,7 +44,7 @@ import System.IO.Unsafe        (unsafePerformIO)
 import System.TimeIt           (timeItT)
 
 import PostgREST.AppState      (AppState, getConfig, getJwtCacheState,
-                                getTime)
+                                getObserver, getTime)
 import PostgREST.Auth.JwtCache (lookupJwtCache)
 import PostgREST.Auth.Types    (AuthResult (..))
 import PostgREST.Config        (AppConfig (..), FilterExp (..),
@@ -161,6 +161,7 @@ middleware appState app req respond = do
   let token  = Wai.extractBearerAuth =<< lookup HTTP.hAuthorization (Wai.requestHeaders req)
       parseJwt = runExceptT $ parseToken conf token time >>= parseClaims conf
       jwtCacheState = getJwtCacheState appState
+      observer = getObserver appState
 
 -- If ServerTimingEnabled -> calculate JWT validation time
 -- If JwtCacheMaxLifetime -> cache JWT validation result
@@ -171,7 +172,7 @@ middleware appState app req respond = do
 
     (True, maxLifetime)  -> do
           (dur, authResult) <- timeItT $ case token of
-            Just tkn -> lookupJwtCache jwtCacheState tkn maxLifetime parseJwt time
+            Just tkn -> lookupJwtCache jwtCacheState tkn maxLifetime parseJwt time observer
             Nothing  -> parseJwt
           return $ req { Wai.vault = Wai.vault req & Vault.insert authResultKey authResult & Vault.insert jwtDurKey dur }
 
@@ -181,7 +182,7 @@ middleware appState app req respond = do
 
     (False, maxLifetime) -> do
           authResult <- case token of
-            Just tkn -> lookupJwtCache jwtCacheState tkn maxLifetime parseJwt time
+            Just tkn -> lookupJwtCache jwtCacheState tkn maxLifetime parseJwt time observer
             Nothing -> parseJwt
           return $ req { Wai.vault = Wai.vault req & Vault.insert authResultKey authResult }