+++ test/io/test_io.py
@@ -1719,6 +1719,8 @@ def test_admin_metrics(defaultenv):
         assert "pgrst_db_pool_available" in response.text
         assert "pgrst_db_pool_timeouts_total" in response.text
 
+        assert "pgrst_jwt_cache_size_bytes" in response.text
+
 
 def test_schema_cache_startup_load_with_in_db_config(defaultenv, metapostgrest):
     "verify that the Schema Cache loads correctly at startup, using the in-db `pgrst.db_schemas` config"
@@ -1737,6 +1739,7 @@ def test_schema_cache_startup_load_with_in_db_config(defaultenv, metapostgrest):
     assert response.status_code == 204
 
 
+# TODO: Rewrite this test using jwt cache size metric
 def test_jwt_cache_purges_expired_entries(defaultenv):
     "test expired cache entries are purged on cache miss"
 
@@ -1830,3 +1833,36 @@ def test_log_pool_req_observation(level, defaultenv):
         else:
             output = postgrest.read_stdout(nlines=4)
             assert len(output) == 0
+
+
+def test_jwt_cache_size_bytes_update_log(defaultenv):
+    "JWT cache size should update on a cache miss"
+
+    env = {
+        **defaultenv,
+        "PGRST_LOG_LEVEL": "info",
+        "PGRST_JWT_CACHE_MAX_LIFETIME": "86400",
+        "PGRST_JWT_SECRET": SECRET,
+    }
+
+    headers = jwtauthheader({"role": "postgrest_test_author"}, SECRET)
+
+    with run(env=env) as postgrest:
+        response = postgrest.session.get("/authors_only", headers=headers)
+        assert response.status_code == 200
+
+        output = sorted(postgrest.read_stdout(nlines=2))
+
+        response = postgrest.admin.get("/metrics")
+        assert response.status_code == 200
+
+        # read cache size from metrics
+        cache_size = float(
+            re.search(r"pgrst_jwt_cache_size_bytes (\d+)", response.text).group(1)
+        )
+
+        assert int(cache_size) > 0
+        assert (
+            "The JWT Cache size updated to " + str(int(cache_size)) + " bytes"
+            in output[1]
+        )